Removed the paasta_serviceinit and marathon_serviceinit code paths
