+++ mypy.ini
@@ -40,9 +40,6 @@ disallow_untyped_defs = True
 [mypy-paasta_tools.marathon_tools]
 disallow_untyped_defs = True
 
-[mypy-paasta_tools.paasta_serviceinit]
-disallow_untyped_defs = True
-
 [mypy-paasta_tools.setup_kubernetes_job]
 disallow_untyped_defs = True
