+++ paasta_tools/api/views/instance.py
@@ -78,7 +78,6 @@ from paasta_tools.mesos_tools import get_tasks_from_app_id
 from paasta_tools.mesos_tools import results_or_unknown
 from paasta_tools.mesos_tools import select_tasks_by_id
 from paasta_tools.mesos_tools import TaskNotFound
-from paasta_tools.paasta_serviceinit import get_deployment_version
 from paasta_tools.smartstack_tools import backend_is_up
 from paasta_tools.smartstack_tools import get_backends
 from paasta_tools.smartstack_tools import HaproxyBackend
@@ -1081,3 +1080,10 @@ def get_marathon_dashboard_links(marathon_clients, system_paasta_config):
     if isinstance(links, list) and len(links) >= len(marathon_clients.current):
         return {client: url for client, url in zip(marathon_clients.current, links)}
     return None
+
+
+def get_deployment_version(
+    actual_deployments: Mapping[str, str], cluster: str, instance: str
+) -> Optional[str]:
+    key = ".".join((cluster, instance))
+    return actual_deployments[key][:8] if key in actual_deployments else None
